losse mapper toegevoegd en deze gebruikt in de controller - de request mappings werken allemaal behalve PUT
