+++ src/main/resources/application.properties
@@ -15,6 +15,10 @@ spring.jpa.database=postgresql
 ## laat de sql commands zien
 spring.jpa.show-sql=true
 
+# database initialization with data.sql after hibernate
+spring.sql.init.mode=always
+spring.jpa.defer-datasource-initialization=true
+
 #spring.datasource.driver-class-name=org.postgresql.Driver
 #spring.jpa.database=postgresql
 #spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
@@ -23,5 +27,3 @@ spring.jpa.show-sql=true
 #spring.jpa.generate-ddl=true
 #spring.jpa.hibernate.ddl-auto=create
 #spring.jpa.show-sql=true
-#spring.sql.init.mode=always
-#spring.jpa.defer-datasource-initialization=true
